Moved stuff from MainActivity to new SearchActivity, test ui updates
