+++ mobile/app/build.gradle
@@ -37,7 +37,7 @@ android {
 }
 
 dependencies {
-
+    implementation 'dev.esnault.wanakana:wanakana-android:1.1.1'
     implementation 'com.google.code.gson:gson:2.8.9'
     implementation 'androidx.core:core-ktx:1.7.0'
     implementation 'androidx.appcompat:appcompat:1.6.1'
