+++ mobile/app/src/main/AndroidManifest.xml
@@ -15,6 +15,13 @@
         android:theme="@style/Theme.DictionaryApp"
         tools:targetApi="31">
         <activity
+            android:name=".SearchActivity"
+            android:exported="false">
+            <meta-data
+                android:name="android.app.lib_name"
+                android:value="" />
+        </activity>
+        <activity
             android:name=".MainActivity"
             android:exported="true">
             <intent-filter>
